api/requirements.txt
fastapi
uvicorn
python-socketio
TikTokLive
python-dotenv